Wizard: Move `PopoverActivation` and `ManageKeysButton`
This moves `PopoverActivation` and `ManageKeysButton` outside from `ActivationKeysList` to their own separate components.
